[enhancement] On demand event loading and conditional scroll
Due to the variable size of the events collection which can, and
generally will, be quite large, loading them every time job details are
displayed is overkill.

Auto scrolling should only happen when the user is not actively using
the content, so the automatic scroll should only be applied if the
bottom of the content box is the current position.

.